+++ src/babyAdd.ts
@@ -102,10 +102,10 @@ export function babyAdd(player: EntityPlayer): void {
 
   // Check if this is a trinket baby.
   if (baby.trinket !== undefined) {
+    const { trinket } = baby;
     const num = baby.num ?? 1;
     repeat(num, () => {
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      smeltTrinket(player, baby.trinket!);
+      smeltTrinket(player, trinket);
     });
     g.itemPool.RemoveTrinket(baby.trinket);
   }