+++ src/classes/babyClasses.ts
@@ -183,7 +183,6 @@ export { LightsBaby } from "./babies/LightsBaby";
 export { LilBaby } from "./babies/LilBaby";
 export { LipstickBaby } from "./babies/LipstickBaby";
 export { LittleHornBaby } from "./babies/LittleHornBaby";
-export { LocustBaby } from "./babies/LocustBaby";
 export { LongBaby } from "./babies/LongBaby";
 export { LostBaby } from "./babies/LostBaby";
 export { LostBlackBaby } from "./babies/LostBlackBaby";